Rename build3 package to builder
