--- build3/.gitignore
@@ -1,4 +1,0 @@
-/.pytest_cache/
-/local/
-__pycache__/
-*.py[cod]
