--- build3/pyproject.toml
@@ -1,30 +1,0 @@
-[build-system]
-requires = ["setuptools>=68"]
-build-backend = "setuptools.build_meta"
-
-[project]
-name = "neurocontainers-build3"
-version = "0.1.0"
-description = "Isolated Python prototype for NeuroContainers Dockerfile generation"
-readme = "README.md"
-requires-python = ">=3.10"
-dependencies = [
-  "jinja2>=3.1",
-  "pyyaml>=6.0",
-]
-
-[project.optional-dependencies]
-test = [
-  "pytest>=8.0",
-]
-
-[project.scripts]
-build3 = "build3.cli:main"
-
-[tool.setuptools.packages.find]
-where = ["src"]
-
-[tool.pytest.ini_options]
-markers = [
-  "parity: exhaustive compatibility checks against the current builder",
-]
