--- build3/src/build3/__init__.py
@@ -1,3 +1,0 @@
-"""NeuroContainers build3 prototype."""
-
-__version__ = "0.1.0"
