+++ builder/__main__.py
@@ -2,4 +2,4 @@ from .cli import main
 
 
 if __name__ == "__main__":
-    main()
+    raise SystemExit(main())
