+++ builder/templates/afni.yaml
