+++ builder/templates/minc.yaml
