+++ builder/templates/miniconda.yaml
