+++ builder/templates/mrtrix3.yaml
