+++ builder/templates/spm12.yaml
