+++ builder/tests/fixtures/file_variants/copied.txt
