+++ builder/tests/test_parity_current_contract.py
@@ -2,10 +2,10 @@ from __future__ import annotations
 
 import json
 
-from build3.config import default_config, resolve_recipe
-from build3.dockerfile import render_dockerfile
-from build3.recipe import compile_recipe
-from build3.release import release_data
+from builder.config import default_config, resolve_recipe
+from builder.dockerfile import render_dockerfile
+from builder.recipe import compile_recipe
+from builder.release import release_data
 
 
 def test_dcm2niix_release_contract_matches_existing_metadata_keys() -> None:
