+++ recipes/sodiumnufft/build.yaml
@@ -77,7 +77,7 @@ readme: |-
   Build the recipe with:
   ```bash
   source env/bin/activate
-  python -m build3 generate sodiumnufft --recreate --architecture x86_64
+  python -m builder generate sodiumnufft --recreate --architecture x86_64
   ```
 
   Full build:
